refactor: move ChainConfig to types crate
